+++ Editor/Components/Light/YPipelineLightEditor.cs
@@ -1,6 +1,7 @@
 ﻿using UnityEngine;
-using UnityEditor;
 using UnityEngine.Rendering;
+using UnityEditor;
+using UnityEditor.Rendering;
 
 namespace YPipeline.Editor
 {
@@ -9,34 +10,71 @@ namespace YPipeline.Editor
     [SupportedOnRenderPipeline(typeof(YRenderPipelineAsset))]
     public class YPipelineLightEditor : LightEditor
     {
-        public Light Light => target as Light;
+        private Light Light => target as Light;
         private YPipelineLight m_YPipelineLight;
+        private SerializedYPipelineLight m_SerializedLight;
+        
 
-        protected new void OnEnable()
+        protected override void OnEnable()
         {
-            base.OnEnable();
             m_YPipelineLight = Light.GetYPipelineLight();
+            m_SerializedLight = new SerializedYPipelineLight(serializedObject, settings);
+        }
+        
+        public void OnDisable()
+        {
+            
         }
 
         public override void OnInspectorGUI()
         {
-            base.OnInspectorGUI();
-            DrawInnerAndOuterSpotAngle();
+            m_SerializedLight.Update();
+            
+            YPipelineLightUI.Draw(m_SerializedLight, this);
+            
+            m_SerializedLight.ApplyModifiedProperties();
         }
-
-        private void DrawInnerAndOuterSpotAngle()
+        
+        protected override void OnSceneGUI()
         {
-            if (!settings.lightType.hasMultipleDifferentValues &&
-                (LightType)settings.lightType.enumValueIndex == LightType.Spot)
+            if (!(target is Light light) || light == null)  return;
+
+            switch (light.type)
             {
-                settings.DrawInnerAndOuterSpotAngle();
-                settings.ApplyModifiedProperties();
+                case LightType.Directional:
+                    using (new Handles.DrawingScope(Matrix4x4.TRS(light.transform.position, light.transform.rotation, Vector3.one)))
+                    {
+                        CoreLightEditorUtilities.DrawDirectionalLightGizmo(light);
+                    }
+                    break;
+                case LightType.Spot:
+                    using (new Handles.DrawingScope(Matrix4x4.TRS(light.transform.position, light.transform.rotation, Vector3.one)))
+                    {
+                        CoreLightEditorUtilities.DrawSpotLightGizmo(light);
+                    }
+                    break;
+                case LightType.Point:
+                    using (new Handles.DrawingScope(Matrix4x4.TRS(light.transform.position, Quaternion.identity, Vector3.one)))
+                    {
+                        CoreLightEditorUtilities.DrawPointLightGizmo(light);
+                    }
+                    break;
+                case LightType.Rectangle:
+                    using (new Handles.DrawingScope(Matrix4x4.TRS(light.transform.position, light.transform.rotation, Vector3.one)))
+                    {
+                        CoreLightEditorUtilities.DrawRectangleLightGizmo(light);
+                    }
+                    break;
+                case LightType.Disc:
+                    using (new Handles.DrawingScope(Matrix4x4.TRS(light.transform.position, light.transform.rotation, Vector3.one)))
+                    {
+                        CoreLightEditorUtilities.DrawDiscLightGizmo(light);
+                    }
+                    break;
+                default:
+                    base.OnSceneGUI();
+                    break;
             }
-        }
-
-        private void DrawYPipelineLight()
-        {
-
         }
     }
 }