+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 
+## [0.1.1] - 2026-02-11
+
+This version is compatible with Unity 6000.3
+
+### Changed
+- Improved Light Component Editor GUI.
+
 ## [0.1.0] - 2026-02-10
 
 This version is compatible with Unity 6000.3
 
 ### Added
-- Per-pixel reflection probe, including parallax correction, normalization & blending.
-- Render Pipeline Asset UI.
+- Added Per-pixel reflection probe, including parallax correction, normalization & blending.
+- Added Render Pipeline Asset Editor GUI.